procedure/codes/requirements.txt
pytorch >= 1.0.0
numpy >= 1.15.4 #For data preparation process
scikit-learn >= 0.20.2 #For AUC-PR